+++ fedlab/utils/dataset/partition.py
@@ -135,6 +135,7 @@ class CIFAR10Partitioner(DataPartitioner):
         self.client_dict = self._perform_partition()
         # get sample number count for each client
         self.client_sample_count = F.samples_num_count(self.client_dict, self.num_clients)
+        self.stats_report = F.partition_report(targets, self.client_dict, class_num=self.num_classes, verbose=False)
 
     def _perform_partition(self):
         if self.balance is None:
