+++ fedlab/utils/functional.py
@@ -147,100 +147,100 @@ def get_best_gpu():
     return torch.device("cuda:%d" % (best_device_index))
 
 
-def partition_report(targets, data_indices, class_num=None, verbose=True, file=None):
-    """Generate data partition report for clients in ``data_indices``.
+# def partition_report(targets, data_indices, class_num=None, verbose=True, file=None):
+#     """Generate data partition report for clients in ``data_indices``.
 
-    Generate data partition report for each client according to ``data_indices``, including
-    ratio of each class and dataset size in current client. Report can be printed in screen or into
-    file. The output format is comma-separated values which can be read by :func:`pandas.read_csv`
-    or :func:`csv.reader`.
+#     Generate data partition report for each client according to ``data_indices``, including
+#     ratio of each class and dataset size in current client. Report can be printed in screen or into
+#     file. The output format is comma-separated values which can be read by :func:`pandas.read_csv`
+#     or :func:`csv.reader`.
 
-    Args:
-        targets (list or numpy.ndarray): Targets for all data samples, with each element is in range of ``0`` to ``class_num-1``.
-        data_indices (dict): Dict of ``client_id: [data indices]``.
-        class_num (int, optional): Total number of classes. If set to ``None``, then ``class_num = max(targets) + 1``.
-        verbose (bool, optional): Whether print data partition report in screen. Default as ``True``.
-        file (str, optional): Output file name of data partition report. If ``None``, then no output in file. Default as ``None``.
+#     Args:
+#         targets (list or numpy.ndarray): Targets for all data samples, with each element is in range of ``0`` to ``class_num-1``.
+#         data_indices (dict): Dict of ``client_id: [data indices]``.
+#         class_num (int, optional): Total number of classes. If set to ``None``, then ``class_num = max(targets) + 1``.
+#         verbose (bool, optional): Whether print data partition report in screen. Default as ``True``.
+#         file (str, optional): Output file name of data partition report. If ``None``, then no output in file. Default as ``None``.
 
-    Examples:
-        First generate synthetic data labels and data partition to obtain ``data_indices``
-        (``{ client_id: sample indices}``):
+#     Examples:
+#         First generate synthetic data labels and data partition to obtain ``data_indices``
+#         (``{ client_id: sample indices}``):
 
-        >>> sample_num = 15
-        >>> class_num = 4
-        >>> clients_num = 3
-        >>> num_per_client = int(sample_num/clients_num)
-        >>> labels = np.random.randint(class_num, size=sample_num)  # generate 15 labels, each label is 0 to 3
-        >>> rand_per = np.random.permutation(sample_num)
-        >>> # partition synthetic data into 3 clients
-        >>> data_indices = {0: rand_per[0:num_per_client],
-        ...                 1: rand_per[num_per_client:num_per_client*2],
-        ...                 2: rand_per[num_per_client*2:num_per_client*3]}
+#         >>> sample_num = 15
+#         >>> class_num = 4
+#         >>> clients_num = 3
+#         >>> num_per_client = int(sample_num/clients_num)
+#         >>> labels = np.random.randint(class_num, size=sample_num)  # generate 15 labels, each label is 0 to 3
+#         >>> rand_per = np.random.permutation(sample_num)
+#         >>> # partition synthetic data into 3 clients
+#         >>> data_indices = {0: rand_per[0:num_per_client],
+#         ...                 1: rand_per[num_per_client:num_per_client*2],
+#         ...                 2: rand_per[num_per_client*2:num_per_client*3]}
 
-        Check ``data_indices`` may look like:
+#         Check ``data_indices`` may look like:
 
-        >>> data_indices
-        {0: array([8, 6, 5, 7, 2]),
-         1: array([ 3, 10, 14,  4,  1]),
-         2: array([13,  9, 12, 11,  0])}
+#         >>> data_indices
+#         {0: array([8, 6, 5, 7, 2]),
+#          1: array([ 3, 10, 14,  4,  1]),
+#          2: array([13,  9, 12, 11,  0])}
 
-        Now generate partition report for each client and each class:
+#         Now generate partition report for each client and each class:
 
-        >>> partition_report(labels, data_indices, class_num=class_num, verbose=True, file=None)
-        Class frequencies:
-        client,class0,class1,class2,class3,Amount
-        Client   0,0.200,0.00,0.200,0.600,5
-        Client   1,0.400,0.200,0.200,0.200,5
-        Client   2,0.00,0.400,0.400,0.200,5
+#         >>> partition_report(labels, data_indices, class_num=class_num, verbose=True, file=None)
+#         Class frequencies:
+#         client,class0,class1,class2,class3,Amount
+#         Client   0,0.200,0.00,0.200,0.600,5
+#         Client   1,0.400,0.200,0.200,0.200,5
+#         Client   2,0.00,0.400,0.400,0.200,5
 
-    """
-    if not verbose and file is None:
-        print("No partition report generated")
-        return
+#     """
+#     if not verbose and file is None:
+#         print("No partition report generated")
+#         return
 
-    if not isinstance(targets, np.ndarray):
-        targets = np.array(targets)
+#     if not isinstance(targets, np.ndarray):
+#         targets = np.array(targets)
 
-    if not class_num:
-        class_num = max(targets) + 1
+#     if not class_num:
+#         class_num = max(targets) + 1
 
-    sorted_cid = sorted(data_indices.keys())  # sort client id in ascending order
+#     sorted_cid = sorted(data_indices.keys())  # sort client id in ascending order
 
-    header_line = "Class frequencies:"
-    col_name = "client," + ",".join([f"class{i}" for i in range(class_num)]) + ",Amount"
+#     header_line = "Class frequencies:"
+#     col_name = "client," + ",".join([f"class{i}" for i in range(class_num)]) + ",Amount"
 
-    if verbose:
-        print(header_line)
-        print(col_name)
-    if file is not None:
-        reports = [header_line, col_name]
-    else:
-        reports = None
+#     if verbose:
+#         print(header_line)
+#         print(col_name)
+#     if file is not None:
+#         reports = [header_line, col_name]
+#     else:
+#         reports = None
 
-    for client_id in sorted_cid:
-        indices = data_indices[client_id]
-        client_targets = targets[indices]
-        client_sample_num = len(indices)  # total number of samples of current client
-        client_target_cnt = Counter(client_targets)  # { cls1: num1, cls2: num2, ... }
+#     for client_id in sorted_cid:
+#         indices = data_indices[client_id]
+#         client_targets = targets[indices]
+#         client_sample_num = len(indices)  # total number of samples of current client
+#         client_target_cnt = Counter(client_targets)  # { cls1: num1, cls2: num2, ... }
 
-        report_line = (
-            f"Client {client_id:3d},"
-            + ",".join(
-                [
-                    f"{client_target_cnt[cls] / client_sample_num:.3f}"
-                    if cls in client_target_cnt
-                    else "0.00"
-                    for cls in range(class_num)
-                ]
-            )
-            + f",{client_sample_num}"
-        )
-        if verbose:
-            print(report_line)
-        if file is not None:
-            reports.append(report_line)
+#         report_line = (
+#             f"Client {client_id:3d},"
+#             + ",".join(
+#                 [
+#                     f"{client_target_cnt[cls] / client_sample_num:.3f}"
+#                     if cls in client_target_cnt
+#                     else "0.00"
+#                     for cls in range(class_num)
+#                 ]
+#             )
+#             + f",{client_sample_num}"
+#         )
+#         if verbose:
+#             print(report_line)
+#         if file is not None:
+#             reports.append(report_line)
 
-    if file is not None:
-        fh = open(file, "w")
-        fh.write("\n".join(reports))
-        fh.close()
+#     if file is not None:
+#         fh = open(file, "w")
+#         fh.write("\n".join(reports))
+#         fh.close()
